uhd: rfnoc: Allow disabling RX radio timestamps
This features enables a setting that can disable timestamps for packets
leaving the RX radio block. By default, timestamps are enabled; with
this change, they can now also be disabled with native GNU Radio APIs,
and from GRC.

The most common use case of this is to enable a direct loopback:

RX Radio -> TX Radio

Because the TX radio interprets timestamps as an instruction when to
send samples, but the RX radio uses timestamps to declare when they were
received, all data packets would be late in this flow graph with
timestamps enabled.

Signed-off-by: Martin Braun <martin.braun@ettus.com>
(cherry picked from commit 55171f63013bc69b9dede093769f59363e586477)
Signed-off-by: Jeff Long <willcode4@gmail.com>