uhd: rfnoc: graph: Correctly name argument 'is_back_edge'
This argument was called 'skip_property_propagation' in the past, which
is not a good name. In UHD, the same change was made. Because the Python
API did not name this argument, this is not even an ABI change, merely
a documentation change.

Signed-off-by: Martin Braun <martin.braun@ettus.com>
(cherry picked from commit 404f014526d0aacfb81a68070e62e05e5bd193b4)
Signed-off-by: Jeff Long <willcode4@gmail.com>