qtgui: Partially revert #6645 change of returnPressed->editingFinished
For the edit_box_msg in particular, returnPressed behavior is desired
so that messages are not sent just when focus is lost. In fact, this was
already changed from editingFinished to ensure this particular behavior.
I have added a comment so that it won't be changed back inadvertently.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>
(cherry picked from commit b5b4fca72f16459656990267fe0cd422e1e647e9)
Signed-off-by: Jeff Long <willcode4@gmail.com>