grc: Add edge (connection) properties
This allows connections (graph edges) to have properties associated with
them. In order to do so, the edge types must have parameters associated
with them in the *.domain.yml file.

For edges that have properties/parameters associated with them,
a property dialog can be summoned by right-clicking a connector arrow.

Signed-off-by: Martin Braun <martin.braun@ettus.com>