grc: Consistently format filenames in messages
GRC usually formats filenames by putting double quotes around them,
but when a flow graph is generated, the filename is formatted as a
Python string, including escaping. On Windows, backslashes become
double backslashes, making the filename harder to read. Here I've
changed the messages to format the filename the same as elsewhere.

Signed-off-by: Clayton Smith <argilo@gmail.com>