runtime: Handle partially done flows
When a block has several outputs or when a block has an output that is
connected to several blocks, if one of the reading blocks is done then
the writing block is considered as done. This commit allows to specify
when a writing block should only be stopped when all its readers are
done. The former behavior is kept as the default one.

Signed-off-by: GRAUX Pierre <pierre.graux@univ-lille.fr>