fec: add FEC_API to CCSDS Reed-Solomon functions
This solves #4454. The (en|de)code_rs_(char|8) functions are
marked as FEC_API so that they can be used by OOT modules. The
(en|de)code_rs_ccsds functions should also be marked as FEC_API.

This is related to #3821. When building with MSVC, the compiler
gives an error if FEC_API is attached to a function definition.
It should be attached to function declarations only.

I have done some clean up to mark all this functions as FEC_API
consistently. They are declared with FEC_API in
include/gnuradio/fec/rs.h, but the implementation was not including
this file, so I have included it now in all the relevant places.
This allows to remove some redundant declarations from char.h
and fixed.h.

I have also added include guards in include/gnuradio/fec/rs.h.
The .h files in reed-solomon/ don't have include guards, but they
are only included in .c files in that folder (to do some sort of
template instantiation in C using the preprocessor).

Signed-off-by: Daniel Estévez <daniel@destevez.net>