fec: use volk_32f_s32f_x2_convert_8u in async_decoder
When Volk >= 3.1 is available, use the new Volk conversion kernel.
This was marked as a TODO when the kernel was added to Volk. It can
be done now, since the kernel has been released in a stable version
of Volk.

Signed-off-by: Daniel Estévez <daniel@destevez.net>
(cherry picked from commit fe87e49b0122ed8d63a06499512335abdf6bd69e)
Signed-off-by: Jeff Long <willcode4@gmail.com>