blocks: add vlen, scale and bias to float_to_uchar
This adds vlen, scale and bias parameters to float_to_uchar. The
vlen and scale parameters are available in float_to_char, but they
were not available in float_to_uchar. The bias parameter is good
to have in a float_to_uchar conversion, since it is often necessary
to apply an additive bias as part of the conversion.

The new Volk kernel volk_32f_s32f_x2_convert_8u(), which is available
in Volk >= 3.1 is used to implement float_to_uchar when building
against a suitable version of Volk. If the Volk version is older,
then there is a ad-hoc implementation in the work() function.

A new QA test for float_to_char using a scale and bias has been added.

The float_array_to_uchar() has also been modified to use
volk_32f_s32f_x2_convert_8u() when available. Even though this kernel
is performing more calculations (multiplication and addition), it is
potentially faster than the plain C implementation, because it uses
SIMD. The float_array_to_uchar() function now seems unused in the
code base, but it is exported in the API.

Signed-off-by: Daniel Estévez <daniel@destevez.net>