grc: Fix compilation of bus ports
This removes a superfluous call to Connection.__init__() in the top
block generator which uses itself as a parent, instead of the original
flowgraph as elsewhere in the same file. 0fd504e5 changed the Connection
object such that passing in a parent that is not an Element is no longer
valid.

Signed-off-by: Martin Braun <martin.braun@ettus.com>
(cherry picked from commit 2d5f73e7d188679e92d81e77ea9004237be57cc7)
Signed-off-by: Jeff Long <willcode4@gmail.com>