grcqt: Improve logging
grc has the comman line option --log, which can be set to debug.
Doing so enables the debug log in Application.py but not in the depending
classes.
For debugging the output of the called classes  would be important,too.

In gtk-grc we have  similar problem. See #4628.

I ported the solution proposed in #4628 to th gui_qt part and it works well.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>