blocks/blockinterleaving: rework permutation check to suppress warning
This was noticed because of what is probably a stdlib bug that warns
about the result of [[nodiscard]] std::unique being discarded.
(std::unique has no business being nodiscard, imho.)

Looked at the code, it does a sort on a permutation vector just to check
whether it's unique and covers $\{0, …, N-1\}$, and then finally, it does
another sort to construct the deinterleaver taps.

Since this is mathematically simple, the deinterleaver tap at position i
simply being the interleaver position of i, and vice versa, the check
for uniqueness and the deinterleaver construction can be done in one
fell swoop. No sorting required!

Made the tests seeded, so that we don't get flaky tests when the random
generator asked to permute $(0, …, 15)$ happens to emit the first $K$
sorted, i.e., $(0, 1, …, K-1 , ×, …)$, which it does in $\frac{1}{16}
\cdot  \frac{1}{15} … \frac{N-K} = \frac{K!}{N!}$ cases.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>
(cherry picked from commit 65e4e9626bacc2c6443f91047fb9d53e416547ce)
Signed-off-by: Jeff Long <willcode4@gmail.com>