Avoid gratuitous const dropping
Don't remove const-ness where there is no need to do so, and use
const_cast when it is necessary. This significantly reduces -Wcast-qual
warnings.

Also, in jack_source, don't add const-ness when not necessary only to
have to cast it away.

Note that the changes in qtgui are incomplete. In particular, this does
not introduce any const_casts in qtgui, nor does it attempt to address
dropping const from point data. The former are a bit more questionable,
while the latter requires more invasive refactoring; accordingly, each
will be addressed in separate commits.

Signed-off-by: Matthew Woehlke <matthew.woehlke@kitware.com>