blocks: add burst to stream block
This adds a new block called Burst to Stream. The block transforms
a bursty tagged stream into a continuous stream by inserting zeros in
the output between packets whenever no packets are available at the
input.

There are two main use cases for this block:

- Using a packet transmitter that outputs a tagged stream with an
  SDR that does not support burst-mode, by converting the output
  of the packet transmitter into a continuous stream of samples
  at the rate required by the SDR.

- Simulation of a packet transmitter by inserting gaps between
  the packets. In this way, a receiver used in the simulation will
  see these packets with gaps between them rather than back-to-back.

An example flowgraph to demonstrate the usage of this block is included.

Signed-off-by: Daniel Estévez <daniel@destevez.net>