modtool/bind: GCC12/13 on Ubuntu needs a flag for binding (#7408)
* modtool/bind: GCC12/13 on Ubuntu needs a flag for binding

There seems to be a bug where on some platforms, GCC fails because it finds the
_GLIBCXX_OPERATOR_DELETE and considers it a "non-usual deallocation function".
Even if that is inherent to C++17. Anyways, in
https://github.com/gnuradio/gnuradio/issues/6477 user boatbod convinces us that
adding -fsized_deallocations helps this. Neither we nor the GCC mailing lists
are perfectly sure why that is, but it does help. Since this is just for binding,
none of the thus-generated code ends up anywhere, which reduces the
headaches of introducing any random compiler flags.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>
Co-authored-by: Martin Braun <mail@martin-braun.org>
(cherry picked from commit 0da3d83d0bf30b133b75990f1c492f51f03c4786)
Signed-off-by: Jeff Long <willcode4@gmail.com>