gui_prefs_file does not need to use os.path.join
When os.path.join is used, it appends an extra directory to the config file path.  For
grc the config file is $HOME/.config/gnuradio/grc.conf/grc.conf and for grc qt the file is
$HOME/.config/gnuradio/grc.conf/grc_qt.conf.

This patch proposes to change this and remove the grc.conf directory from the path.  This would
make the grc config file $HOME/.config/gnuradio/grc.conf and the grc qt config file
$HOME/.config/gnuradio/grc_qt.conf.

All that is needed to accomplish this is to remove the use of os.path.join to generate the
gui_prefs_file.

Signed-off-by: Chris Gorman <chrisjohgorman@gmail.com>