grc: Enable pluggable workflow
In the previous version of GRC, workflows were hard-coded. There was C++
vs.  Python, with and without Qt, as hier block or as top block. This
changeset enables a different paradigm: All workflows are now available
as plugins. By default, GNU Radio supports all workflows that were
previously hardcoded (see list above, plus gr-bokeh_gui with Python).

A major part of this changeset is the updated Options block. The YAML
for the Options block no longer lists all the available workflows.
Rather, there is a Python-based implementation of the Options block,
which will dynamically load all workflows it can find.

Co-authored-by: Martin Braun <martin.braun@ettus.com>
Signed-off-by: ZakyHermawan <zaky.hermawan9615@gmail.com>
Signed-off-by: Martin Braun <martin.braun@ettus.com>