grc-qt: Fix failing tests
As the behaviour of the blocklib window was modified, the corresponding tests must be modified , too.
The first idea is to insert a KeyDown sequence in the add_block_from_query.
But that did'nt work for me. The KeyDown sequence was not executed.
The workaround is to use the press function rom pyautogui. To work properly some delays must be inserted.

Fixes: #7448

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>