runtime: Use ABC's in tag_utils
Currently, tag_utils detects if an argument in python is an
integer/list/whatever by checking its type. This can lead to
hard to track bugs if e.g. the programmer attempts to use
a numpy np.int_ or an ndarray. These types can be supported
by having isinstance check for an abstract base class.

Signed-off-by: souponaroom <36963544+souponaroom@users.noreply.github.com>