grc: Attribute error
grc --gtk
If you disable a block in flowgraph not all depending blocks are marked as invalid.
Reenabling this block marks the fg as invalid. You have to open and close a block to mark
the fg as valid.

grc -qt
Disableing and reenabling work without errors.
But disableing and Undo'ing leaves a fg invalid.

This issue is fixed by this pr.

Fixes #7367

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>