python: set numpy type propagation policy in gateway
This way, Python code using GNU Radio blocks has the same type
propagation behaviour as numpy 2, no matter whether it's using numpy 1
(>=1.24.0) or 2.

Background is that numpy2 makes test cases fail that do something like

```python
a = numpy.ndarray([1,2], dtype=numpy.uint8)
b = a[0] * 256 + a[1]
```

because numpy2's propagation rules convert 256 to the dtype of `a[0]`
first, and that fails, because 256 is too large to fit an 8 bit integer.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>