gr_modtool: util_functions.py add newline to variable modified_last
I would like someone to confirm that this fix works for linux.  Under
windows this fixes a problem with python modifying __init__.py imporperly
and I believe this will occur under linux as well, but I have not tested
with linux.

I would also like to ensure that this doesn't negatively impact cpp code
generation as I have not yet tested it with cpp.

Signed-off-by: Chris Gorman <chrisjohgorman@gmail.com>