uhd: Update RFNoC image builder GRC bindings
This updates the portion of gr-uhd that deals with generating RFNoC
bitfiles from within GRC. For now, this is limited to generating a .grc
file with GNU Radio Companion that describes an RFNoC image core; this
file must then be passed to `rfnoc_image_builder` with the `-r`
argument, the same way that regular image core YAML files are passed
with the `-y` argument.

The following changes are included:
- All available image core blocks were updated to match recent UHD
- New image core blocks were added (addsub, CHDR DMA transport adapter,
  no-connect, fosphor aka f15, keep-one-in-N, null source/sink,
  replay/capture, stream endpoint, signal generator, split-stream,
  switchboard).
- The X310 BSP block was replaced with a generic BSP block, which
  supports all currently available RFNoC devices.
- The domains were updated, and domains for axi4_mm, gpio, and transport
  adapters were added.
- Two examples for image core files were added: An X3x0 image, which can
  be used to generate X300_HG, X300_XG, X310_HG, and X310_XG standard
  bitfiles, and an x410_X4_200 example.
- A README.md is added in gr-uhd/examples/grc to better explain what's
  going on.

The usage of this to generate RFNoC bitfiles depends on the availability
of updated UHD. However, note that with current UHD and GNU Radio, the
image builder feature of gr-uhd was not functional.

Signed-off-by: Martin Braun <martin.braun@ettus.com>