grc-qt: some improvements on the PropertyDialog
Multiline text cannot be saved  ( In comments)
This is fixed now

The return key is connected to the accept button. So you have to reopen the dialog, to do further changes.
Now this key only finishes the input on an entry and you can modify this entry again, without reopening.

After changing a parameter the gui is now updated.

For instance open a filter block and change the window parameter from Hamming to Kaiser, than the Beta parameter appears.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>