blocks: Fix useless byte size functions in wavfile_sink.
The current wavfile_sink block API has getter and
setter functions for the sample bit size. However,
the setter function's value is ignored because the
sample size is fully determined by the file format
and subformat. Additionally, the code is inconsistent,
as the getter returns bytes instead of bits, and
there is a risk of using uninitialized memory.

This update removes unnecessary operations and
resolves the inconsistency between bits and bytes.

Signed-off-by: Ladislav Sladecek <ladislav.sladecek@gmail.com>