blocks: Reopen wavfile_sink after stop().
Fixes #7456 .

The flowgraph uses the stop() method of a block to
indicate that it has been locked for
configuration. I believe the correct behavior for
file-based sinks is to write all data and headers
to disk, flush the file buffers, but avoid closing
the file (see also #2590).

This behavior is implemented in the
file_sink. Since the file_sink is very simple and
does not require updating headers, flushing the
file is sufficient.

On the other hand, wavfile_sink can write to
several file formats, some of which include
headers. The implementation relies on libsndfile,
which provides a flush function. However, the
source code indicates that this function does not
update headers before flushing.

The current implementation of wavfile_sink calls
the close() method within the stop() method to
ensure the file on disk is correct. However, this
also closes the file.

My solution is to add a flag for automatic
reopening. When the file is opened, its filename
is stored. When the stop() method is called, the
file is closed, and the automatic reopen flag is
set. Later, within the start() method the file is
automatically reopened in append mode.

Signed-off-by: Ladislav Sladecek <ladislav.sladecek@gmail.com>