gr-analog : minor correctly show the two audio tones on a real spectrum.
The example showcases two tones at 400Hz and 350Hz that are added.
However, the total audio spectrum is offering such a low resolution
that the two peaks are not visible.
In fact, two peaks are present on the spectrum but
it is only the negative replica of the real signal.
It could be misleading.
To clarify, a higher resolution allows to distinguish the two peaks.
And because it is float (real values),
there is no need of the negative part of the spectrum.

Signed-off-by: Adrien Michel <adriengit@users.noreply.github.com>