gr-analog: minor, provides a total signal (two sines) with peaks to 1
In FM, the amplitude of the audio signal to be modulated is
directly related to the spectrum.
The relationship is involving the the frequency deviation (75kHz),
a parameter that is standard for broadcast FM.

However, if the input signal is not at 1, but scaled down by a factor 2,
the final channel bandwidth is not the same.

Thanks to this modification, the current example can be used to
showcase the carson rule on the FM bandwidth:
Bw = 2*delta_freq + 2 * M

Signed-off-by: Adrien Michel <adriengit@users.noreply.github.com>