grc/core/blocks/embedded_python: use current block parameter values when extracting io signature
This makes it possible to have correctly working 'vlen' block
parameters in embedded python blocks. Before this change adding such a
block parameter and using it to determine the size of a vector
input/output would ignore the value set in the GUI and unconditionally
use the default argument value from the blocks `__init__()` function.
In principle this should also make it possible to change the datatype
of a block port depending on a block parameter, though of course
without the nice dropdown menu known from the builtin blocks.

Adding this functionality requires only two small changes:

1. The actual parameter values (as set via the GUI parameters dialog)
   need to be passed to the block IO extractor such that the block can
   be instantiated with these values instead of solely the defaults.

2. An early return statement needs to be removed form the embedded
   python block `rewrite()` implementation, this statement would abort
   the rewrite early if the embedded python source code was
   unchanged, even though block parameters might have changed (see
   `Block.import_data()`). Before the change described under (1.) a
   second invocation of `rewrite()` would have always resulted in the
   same effective block "state" even if the block parameters had
   changed, so I presume that this early return was an optimization to
   avoid the redundant second block IO extraction and port setup.
   However with the change described under (1.) the block IO
   extraction now needs to be performed at least twice: first to
   extract the names of the block parameters that should be imported
   and second to instantiate the block with the imported parameter
   values to obtain the correct IO signature.

I believe that this is actually the intended purpose and use of the
"fixed-point iteration" performed by `Block.import_data()`: ensuring
that `rewrite()` is called with the final "stable" set of parameters.
With the early return removed Embedded Python Blocks will
typically (i.e. always) be instantiated twice. If none of the IO
signatures, or other "special" block parameters (such as the name),
depend on any of the Python init arguments the second instantiation is
technically redundant.

Signed-off-by: Johannes Beisswenger <johannes@beisswenger.name>