grc: don't draw block port GUI element without valid dimensions
The `Port` constructor initializes the rectangle area to an empty
list. If subsequently `draw()` is called before `create_shapes()` is
called the (attempted) expansion of this empty list into four
dimension parameters raises an exception. This does not crash grc
however the exceptions appear in the log output. The solution is to
simply not draw anything as long as the rectangle area isn't valid
i.e. until after `create_shapes()` has been called.

There are other alternative fixes which would also be possible:
 - call `Port.create_shapes()` in the `Port` constructor. However this
 could potentially "race" with regards to the port's rotation, i.e. the
 port rectangle might initially be called with an incorrect
 orientation for a few frames. Though after further investigation I
 think the rotation would always be correct, or rather consistent with
 the parent block.
 - ensure this (and other) elements are only added to some "drawlist"
 until they are properly instantiated, i.e. after `create_shapes()`
 has been called. Notably this would require some substantial
 refactoring since the "drawlist" (updated by
 `FlowGraph.update_elements_to_draw()`) is used by the top-level
 `FlowGraph.create_shapes()` and `FlowGraph.create_labels()` functions
  and would essentially need to be duplicated (essentially
  "double-buffered").

Signed-off-by: Johannes Beisswenger <johannes@beisswenger.name>