grc: add test for epy blocks with parameter dependent IO signature
The example GRC file contains an embedded python block taking a `vlen`
parameter that determines the length of its I/O vector ports. The
parameter defaults to value of `vlen=1` while the flow graph
configures all blocks to use vectors of length `64`. This ensures that
the overall flow graph is only valid if the _configured_ value of the
`vlen` parameter is used during I/O signature extraction instead of
the _default_ value.

Signed-off-by: Johannes Beisswenger <johannes@beisswenger.name>