grc/gui: update ports/labels when rewriting block within PropsDialog
When editing block parameters using the `PropsDialog` the block is
rewritten and validated in order to provide updated error messages
about potentially invalid values at the bottom of the dialog window.
Previously these rewrites were not accompanied with matching updates
to the blocks GUI elements such as the port symbols and
labels (`creatr_shapes()` and `create_labels()`), these would only be
updated upon exiting the `PropsDialog`. This could result in "mildly
invalid" graphics in the background, until closing the
`PropsDialog`. ("mildly invalid": when reducing multiplicity of e.g. an
"Add" block, port symbols would be removed from the block while the
`PropsDialog` was still open, but the block dimensions would only be
updated once the dialog was closed)

As a result of changes to the embedded python IO extraction
code (which now allow changes to the block IO signature by changing
block parameters) new ports can now be created as a result of changing
block parameters. This, combined with the (now fixed) lack of calls to
`create_shapes()` and `create_labels()`, could result in calling
`draw()` on not fully constructed GUI Port elements, which would throw
an exception on every main window `draw()` until the `PropsDialog` was
finally closed.

Signed-off-by: Johannes Beisswenger <johannes@beisswenger.name>