runtime: tagged_stream_block: don't removes tags, instead don't show them
Before, the TSB superclass used to modify the contents input buffer tags
vector, adding itself to the list of blocks whom to not show the tag on
tags with the length tag key.

Now, instead of doing that, it simply "filters" what TSBs see by
(non-virtually) overriding get_tags_in_{range,window}.

This removes a contention point, avoids adding complex data to the
oft-copied blocks, and removes the usage of the deprecated
`block::remove_item_tag`. It was the last consumer of that function, so
that can now be retired.

Fixes a small problem with `qa_ofdm_txrx` (1040.0 is not exactly the
same as 1040), and correctly copies tags from in- to output in
`digital::packet_headergenerator_bb_impl`.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>