runtime: rm block::remove_item_tag; rm (tag∊delete list)
With TSBs no longer marking tags as deleted in their input buffer, we
can drop the "Will be removed in 3.8"-deprecated
`block::remove_item_tag`.

With that gone, nothing can put blocks on the "am I deleted for this
particular block?" vector of tags. Hence, the functions called from
`block::get_tags_in_range`, `block_detail::…`, and
`buffer_reader::get_tags_in_range` no longer need to check whether a
block is on that list. Hence, remove that check, and the "who is
asking?" `id` argument.

These were all long-deprecated or internal/detail API anyways, so we're
not breaking compatibility.

It'd be more than desirable to remove the deletion list from `tag_t`,
but that's an API breakage (though I **really** hope nobody was using
that public field of the struct), and should be done in a separate
commit.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>