PMT: Remove detrimental calls to 'pubseekoff' from de-serialization
'pubseekoff' method was used probably with intention to move
the std::streambuff's internal position into the right place.
However this was already done by 'sgetn' method. With both
calls: to 'sgetn' and to 'pubseekoff', the internal position
of the 'sb' streambuff ended up being wrong.

This commit fixes the issue #7672.

Signed-off-by: Piotr Krysik <ptrkrysik@gmail.com>