github CI: upload installation files
Upload an archive of the installed files, so that one can locally run
the compiled GNU Radio exactly as built.

I wanted to also add coredumping, but couldn't, because that would have
required working with the host outside the builder containers.

Github CI would only allow us to set the core pattern, and hence to
collect the coredumps, from a privileged container. Not going that
security nightmare route.

If we wanted that, we'd, instead of directly running each element of the
matrix in a container, run them each on a VM, and manually start docker
from there. Not going that route today.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>