grc-qt: block library follow up to #7644
With #7644 on option was introduced to display the lib in collapsed or expanded state.
But due to a bug only the first category was expanded. This is fixed now.

The search field was not cleared, if you selected a block by key_down and then pressing return.

This is fixed now, too.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>