github CI: use explicit ubuntu-24.04 instead of -latest
a) not a fan of using an unspecified version of software in CI.
b) Github has been spamming the warning with 'will be replaced by the
week of 2025-01-17' for *every single use*,  and then that didn't even
happen. Considering this, we shouldn't put too much trust into that not
being the case for longer, or again soon.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>