uhd: Fix connecting back-edges
gr-uhd contained a bug where back-edges were not being properly
propagated into UHD (meaning, the setting was not being set in the
corresponding UHD call).

In addition, we add a warning when attempting to define an edge to
a streamer node as a back-edge. Because these edges are never the cause
of a loop, the underlying UHD API does not allow tagging these as
back-edges. If someone attempts to declare a streamer-edge as
a back-edge, we will now emit a warning on the logger interface.

Signed-off-by: Martin Braun <martin.braun@ettus.com>