grc-qt: Right-clicking in the canvas doesn't bring up the context menu
Right-clicking in the canvas now brings up a context menu on a selected block.
If multiple blocks are selected the context menu is assoziated to th 'first' block
in the list. If no block is selected a message appears.

If the mouse is over a block this block will be selected.

Fixes #7610

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>