uhd: rfnoc: Remove enum block type from image builder blocks
The image builder GRC bindings all have a 'type' parameter, which is
hidden, and is used by `rfnoc_image_builder` to identify the block type.
It was defined as an enum, to limit (and document) the available
options.

However, in GRC, when you push the down or up keys on the keyboard while
selecting a block, you change the first parameter that is an enum. For
example, if you selected the terminator block and push down, it changes
the 'type' to 'block', which confuses `rfnoc_image_builder`.

The "fix" is to simply make all of these 'type' parameters strings.

Signed-off-by: Martin Braun <martin.braun@ettus.com>