grc: fix wrong path for generated yml file when generating C++ hier block
self.file_path_yml should be pointing to the GRC_HIER_PATH or state directory
so grc can search for the hier block.

With this, when you generate C++ No GUI flow graph, if you have a hier block,
the hier block will also be included in the header file of the top block.

However, if the No GUI flow graph contains hier block, the generated code still can't be run.
We still need to fix the templates and/or the python code that responsible for code generation.

Signed-off-by: Zaky Hermawan <zaky.hermawan9615@gmail.com>
Signed-off-by: Marcus Müller <mmueller@gnuradio.org>