audio: unify priorities between sink and source architectures
I made a mistake 4 years ago that gave the "windows" sink a different
priority than the source. That led to great user confusion.

Make sure that doesn't happen again. Instead, keep priorities in a
central map.

Since the `register_{sink,source}` function was rendered unused by that,
and since it's really not a necessary function, remove it. It's in
lib/audio_registry.h, so can't be used out of tree, anyways, and thus
not API.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>