digital: fix bugs in band-edge filter design
The band-edge filter that is designed in the FLL Band-edge block
has two bugs:

- The width of the frequency domain response of the filter is
  2 * rolloff, but it should be rolloff.

- The filter is translated in frequency to be centered at
  (1 + rolloff) / (2 * sps), but it should be centered at
  1 / (2 * sps).

These problems, and their fixes, are demonstrated in this Jupyter
notebook:

https://github.com/daniestevez/jupyter_notebooks/blob/master/Band-edge%20filter.ipynb

Additionally, this commit changes "Prototype Filter Size" in the
GRC YAML to "Filter Size". The band-edge filter is not a Polyphase
Filterbank, so using the word "Prototype" here can be misleading.

Signed-off-by: Daniel Estévez <daniel@destevez.net>