runtime/qa_get_first_tag_in_range: Don't test until underlying deadlock solved
Argilo demonstrated in https://github.com/gnuradio/gnuradio/issues/7711
pretty concisely what the problem is:

Our C++-wrapping doesn't hold the GIL. When we pass in a python function
to be executed by the predicate-checking functionality, that however
needs the GIL. It does hold the buffer mutex.

When, from another place, holding the GIL, something tries to acquire
the buffer mutex, it blocks until get_first_tag_in_range yields it.
Which it can't, because it is waiting for the GIL to be yielded.
Deadlock.

This is not trivial to fix (it requires reworking the kind of locking
happening – currently, there's but one buffer mutex, protecting all:
you can't even concurrently read tags or the buffer pointers, let alone
write them; and all this happens with no chance of interruption when the
critical sections are through.)

So, solution to unblock GR main development for now is to disable this
QA.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>