CI: don't make Python formatting fails a blocker to unit tests
We don't offer any tooling to do exactly the right formatting in the
IDE.
Having to wait for CI just to see it not even run because your IDE and
pycodestyle (of an old version) disagree on something is not conductive
to code quality, overall.

I'm leaving the check in, and we'll see failures and need people to fix
them. But let's not gate ourselves from progress this much.

Proposal would probably be to actually pick a tool that integrates with
IDEs well, or at least is foolsafe to run manually, and base the check
on that.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>