grc:  Using the Escape key in PropsDialog is buggy
See #7877

But the problem is more general and affects blocks with ports.
Changing the port number (input or output ) in a block and pressing Esc
removes the block area and only the ports can be seen in the flowgraph.

Pressing Esc in the PropsDialog for blocks without ports the key is mapped to accept,
but I think the common use should be 'Abort Editing'

So this pr maps Esc to reject.

Fixes #7877

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>