ci: windows installer: enable ninja tests on bundle installer
Enabled cmake/ninja based testing by adding flag - DENABLE_TESTING=ON
to the cmake command of the GH action. To be able to execute those tests
on the bundle installer the following changes need to be implemented:
- removed dynamic linking: when boost is installed by conan only static
libraries (*.lib, and no *.dlls) are installed, which require static linking.
- use installation python environment for installer testing: by changing python
executable from the bundle installer for QA tests on Windows Installer builds
when GR_BUILD_INSTALLER variable is set.

Signed-off-by: Clemens Felber <clemens.felber@emerson.com>