uhd: Add PDU port to RFNoC Tx streamer
This allows sending PDUs to the RFNoC Tx streamer instead of regular
streams. In addition to the normal streaming port, the RFNoC Tx Streamer
block now also has a message port that accepts PDUs. This can be used
to implement bursty communication, radar bursts, or other things that
are better mapped to PDUs than continuous streams.

Messages can include a timestamp for timed transmit. By default, every
PDU is transmitted with an end-of-burst flag, but that can be disabled
for back-to-back PDUs.

Another example (rfnoc_duc_radio_pdu.grc) is added to show off this
feature with a message strobe.

Signed-off-by: Martin Braun <martin.braun@ettus.com>