TCP- Server: Fix crash on a connection reset package
See issue 7390

To fix this issue, it's necessary to handle the error in
d.socket_shutdown.

I think, it's helpfull for the user to report an info in this case.

So I added a funcion get_logger in basic_block, to get access to the
block's logger.
The example in 7390 now results in
socket_pdu :info: Connection reset by peer
socket_pdu :info: Transport endpoint is not connected
when receiving a reset and the process keeps running.

Fixes #7390

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>