TCP- Server: Fix crash on a connection reset package
See issue 7930

To fix this issue, it's necessary to handle the error in
d.socket_shutdown.

I think, it's helpfull for the user to report an info in this case.

So I configured a default logger

The example in 7930 now results in
./tcpbug.py
Press Enter to quit:
tcp_connection :info: Connection reset by peer
tcp_connection :info: Transport endpoint is not connected

when receiving a reset and the process keeps running.

Fixes #7930

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>